Project3/requirements.txt
# Root requirements.txt - Master list for all components
yfinance
requests
feedparser
python-dotenv
groq
fastapi
uvicorn[standard]
